Review folder structure inside assignment folder. Teacher & Student has separate files. Ready to implement other features.
