+++ frontend/src/components/teacher/TeacherHomePage.jsx
@@ -2,6 +2,7 @@
 import { useRecoilState } from "recoil";
 import { userState } from "../../js/state-information";
 import LecturePage from "../lecture/LecturePage";
+import TeacherAssignmentPage from "../assignment/TeacherAssignmentPage";
 
 
 export default function TeacherHomePage() {
@@ -13,6 +14,7 @@ export default function TeacherHomePage() {
       <h4 className="card-title">Teacher</h4>
       {user[0].name}
       <LecturePage />
+      <TeacherAssignmentPage />
     </div>
   );
 }
